initail commit for mul

+++ 0x0C-more_malloc_free/101-mul.c
@@ -11,7 +11,7 @@
 
 int main(int argc, char const *argv[])
 {
-int n1, n2;
+unsigned long int n1, n2;
 if (argc != 3)
 {
 printf("Error\n");
@@ -19,6 +19,6 @@ exit(98);
 }
 n1 = atoi(argv[1]);
 n2 = atoi(argv[2]);
-printf("%i\n", n1 * n2);
+printf("%lu\n", n1 *n2);
 return (0);
 }
